More basic UI skeleton.
Switched out some functions for the support library versions.

Switched the action bar out for a toolbar which is now used to access the drawer list.

The buttons in the drawer list now function and can swap to the two currently made fragments; the class list and the profile page