Fixed RequestListActivity location issue.
When returning to the request list after viewing a request or making a new one, it would no longer get location updates. This adds a call to start the gps in the onResume() function so it can continue getting location updates.

Also switched the constructor calls to use the application context instead of the activity context. It works the same, but it's a precaution for later n case opening the app through a notification causes an error when closing an activity.