Added check for user group being deleted.
When the user loads the My Group page and has an internet connection the app will check if the request still exists. If it doesn't it will automatically leave the group.

Moved the call to finish() when loading the request info to avoid crashes when there is no network connection.

Added check to prevent pressing the join and leave group buttons multiple times in a row.